convert all tensors to float64 in original autoencoder

+++ models/autoencoder.py
@@ -21,8 +21,8 @@ class Autoencoder(torch.nn.Module):
             for i, data in enumerate(train_dl):
                 optimizer.zero_grad()
                 t_x_point, t_y_point, t_y_mask, t_channel_pow, file_path, j = data
-                t_x_point, t_y_point, t_y_mask = t_x_point.to(torch.float32).to(device), t_y_point.flatten(1).to(device), t_y_mask.flatten(1).to(device)
-                t_channel_pow = t_channel_pow.flatten(1).to(device)
+                t_x_point, t_y_point, t_y_mask = t_x_point.to(torch.float64).to(device), t_y_point.flatten(1).to(torch.float64).to(device), t_y_mask.flatten(1).to(torch.float64).to(device)
+                t_channel_pow = t_channel_pow.flatten(1).to(torch.float64).to(device)
                 t_y_point_pred = self.forward(t_x_point).to(torch.float64)
                 loss_ = torch.nn.functional.mse_loss(t_y_point * t_y_mask, t_y_point_pred * t_y_mask).to(torch.float64)
                 if loss == 'rmse':
